requirement.txt
torch
torchvision
tqdm
matplotlib
click
spacy
Pillow
pycocotools